requirements.txt
django
djangorestframework
mysqlclient
coverage
pytz
pandas
scipy
